fadez 2 fully implemented working

+++ LeftBrain/PedalBrainz_LeftBrain_Software/v2/PedalBrainz_LeftBrain_Neo_Buttons_51_Refactor_01/PedalBrainz_LeftBrain_Neo_Buttons_51_Refactor_01.ino
@@ -266,7 +266,7 @@ float scalings[2][3] {
 
 int brightness = 0;
 int brightnessScaled;
-int fadeAmount = 1;  //Change it to adjust the fading between each step (fading speed)
+float fadeAmount = 1;  //Change it to adjust the fading between each step (fading speed)
 int maxBrightness = 190; //Pixel Max Bright
 int maxBrightnessTemp = 190; //Pixel Max Bright
 int delayAmount = 0;
